update library manager module
